Add ActiveForm Widget
Create class ActiveForm in the application components namespace, to override CActiveForm.
Override the methods label(), labelEx(), error() and errorSummary() to allow form objects to be passed as well as form models.
Add the methods input() and button() to automatically render the correct input for a form element or button. Add the method hint() to render an input element's hint text.